change long to int, make put requests more semantic

+++ rest-services/src/main/java/com/pcalouche/spat/restservices/api/service/TeamService.java
@@ -5,11 +5,11 @@ import com.pcalouche.spat.restservices.api.dto.TeamDto;
 import java.util.List;
 
 public interface TeamService {
-    TeamDto findById(Long id);
+    TeamDto findById(Integer id);
 
     List<TeamDto> findAll();
 
     TeamDto save(TeamDto team);
 
-    void delete(Long id);
+    void delete(Integer id);
 }
